ci(moonrun): isolate async conformance tests
