+++ .github/workflows/moonrun-conformance.yml
@@ -1,0 +1,99 @@
+# moon: The build system and package manager for MoonBit.
+# Copyright (C) 2024 International Digital Economy Academy
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+# For inquiries, you can contact us via e-mail at jichuruanjian@idea.edu.cn.
+
+name: Moonrun async conformance
+
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'crates/moonrun/**'
+      - 'third_party/moonbitlang_async'
+      - 'crates/moon/tests/test_cases/moon_test/async_wasm.rs'
+      - '.github/workflows/moonrun-conformance.yml'
+  pull_request:
+    branches:
+      - main
+    paths:
+      - 'crates/moonrun/**'
+      - 'third_party/moonbitlang_async'
+      - 'crates/moon/tests/test_cases/moon_test/async_wasm.rs'
+      - '.github/workflows/moonrun-conformance.yml'
+  schedule:
+    - cron: '0 20 * * *'
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+permissions:
+  contents: read
+
+jobs:
+  test:
+    name: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os:
+          - ubuntu-latest
+          - macos-latest
+          - windows-latest
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Configure git
+        run: git config --global core.autocrlf false
+
+      - uses: actions/checkout@v4
+        with:
+          submodules: true
+
+      - uses: dtolnay/rust-toolchain@1.95.0
+
+      - name: Cargo cache
+        uses: Swatinem/rust-cache@v2
+
+      - name: Install MoonBit (Unix)
+        if: matrix.os != 'windows-latest'
+        run: |
+          curl -fsSL https://cli.moonbitlang.com/install/unix.sh | bash -s nightly
+          echo "$HOME/.moon/bin" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/target/debug" >> $GITHUB_PATH
+
+      - name: Install MoonBit (Windows)
+        if: matrix.os == 'windows-latest'
+        run: |
+          $env:MOONBIT_INSTALL_VERSION="nightly"
+          Set-ExecutionPolicy RemoteSigned -Scope CurrentUser; irm https://cli.moonbitlang.com/install/powershell.ps1 | iex
+          "$env:USERPROFILE\.moon\bin" | Out-File -FilePath $env:GITHUB_PATH -Append
+          "$env:GITHUB_WORKSPACE\target\debug" | Out-File -FilePath $env:GITHUB_PATH -Append
+
+      - name: Build Moon and Moonrun
+        run: cargo build -p moon -p moonrun
+
+      - name: Bundle core (Unix)
+        if: matrix.os != 'windows-latest'
+        run: cargo run --bin moon -- -C ~/.moon/lib/core bundle --all
+
+      - name: Bundle core (Windows)
+        if: matrix.os == 'windows-latest'
+        run: cargo run --bin moon -- -C "$env:USERPROFILE\.moon\lib\core" bundle --all
+
+      - name: Test Moonrun against upstream async
+        run: cargo test -p moon --test mod test_cases::moon_test::async_wasm::test_async_wasm_upstream -- --ignored --exact
