+++ crates/moon/tests/test_cases/moon_test/mod.rs
@@ -16,6 +16,7 @@
 //
 // For inquiries, you can contact us via e-mail at jichuruanjian@idea.edu.cn.
 
+mod async_wasm;
 mod patch;
 #[cfg(unix)]
 mod use_cc_for_native_release;
@@ -26,114 +27,6 @@ use expect_test::expect_file;
 use crate::dry_run_utils::assert_lines_in_order;
 
 use super::*;
-
-fn repo_root() -> std::path::PathBuf {
-    std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(|path| path.parent())
-        .unwrap()
-        .to_path_buf()
-}
-
-// Upstream async has tick-sensitive tests; keep wasm package runs isolated
-// from the Rust test harness's package-level concurrency.
-static ASYNC_WASM_TEST_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-const MOONBIT_ASYNC_CHECK_FD_LEAK: &str = "MOONBIT_ASYNC_CHECK_FD_LEAK";
-
-fn prepare_async_wasm_workspace(dir: &TestDir) -> std::path::PathBuf {
-    let repo_root = repo_root();
-    let async_dir = repo_root.join("third_party/moonbitlang_async");
-    let async_member = async_dir
-        .to_string_lossy()
-        .replace('\\', "\\\\")
-        .replace('"', "\\\"");
-
-    std::fs::write(
-        dir.join("moon.work"),
-        crate::util::read(dir.join("moon.work.template"))
-            .replace("@@ASYNC_MEMBER@@", &async_member),
-    )
-    .unwrap();
-    std::fs::copy(dir.join("app/moon.mod.template"), dir.join("app/moon.mod")).unwrap();
-
-    async_dir
-}
-
-fn run_async_wasm_package(dir: &TestDir, package: &str) -> String {
-    let _guard = ASYNC_WASM_TEST_LOCK
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    let moonrun = moonrun_bin();
-    let output = moon_cmd(dir)
-        .env("MOON_OVERRIDE", moon_bin())
-        .env("MOONRUN_OVERRIDE", &moonrun)
-        .env(MOONBIT_ASYNC_CHECK_FD_LEAK, "1")
-        .args([
-            "-C",
-            "app/main",
-            "test",
-            "--target",
-            "wasm",
-            "--package",
-            package,
-            "--sort-input",
-        ])
-        .assert()
-        .success()
-        .get_output()
-        .stdout
-        .clone();
-
-    std::str::from_utf8(&output).unwrap().to_owned()
-}
-
-fn run_upstream_async_wasm_package(package: &str) -> String {
-    let _guard = ASYNC_WASM_TEST_LOCK
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    let moonrun = moonrun_bin();
-    let async_dir = repo_root().join("third_party/moonbitlang_async");
-    let build_dir = async_dir.join("_build");
-    std::fs::create_dir_all(&build_dir).expect("failed to create async test build dir");
-    let target_dir = tempfile::Builder::new()
-        .prefix("moon-test-target-")
-        .tempdir_in(&build_dir)
-        .expect("failed to create async test target dir");
-    let moon = moon_bin();
-    let path = std::env::join_paths(
-        std::iter::once(
-            moon.parent()
-                .expect("test moon binary should have a parent directory")
-                .to_path_buf(),
-        )
-        .chain(std::env::split_paths(
-            &std::env::var_os("PATH").unwrap_or_default(),
-        )),
-    )
-    .expect("failed to put the test moon binary on PATH");
-    let output = moon_cmd(&async_dir)
-        .env("MOON_OVERRIDE", &moon)
-        .env("MOONRUN_OVERRIDE", &moonrun)
-        .env(MOONBIT_ASYNC_CHECK_FD_LEAK, "1")
-        .env("PATH", path)
-        .arg("--target-dir")
-        .arg(target_dir.path())
-        .args([
-            "test",
-            "--target",
-            "wasm",
-            "--package",
-            package,
-            "--sort-input",
-        ])
-        .assert()
-        .success()
-        .get_output()
-        .stdout
-        .clone();
-
-    std::str::from_utf8(&output).unwrap().to_owned()
-}
 
 #[test]
 fn test_moon_test_succ() {
@@ -754,153 +647,6 @@ fn test_async_test() {
     );
     let last_line = out2.lines().last().unwrap_or("");
     check(last_line, expect!["Total tests: 1, passed: 0, failed: 1."])
-}
-
-#[test]
-fn test_async_wasm_workspace_timer() {
-    let dir = TestDir::new("moon_test/async_wasm_workspace_timer");
-    prepare_async_wasm_workspace(&dir);
-
-    check(
-        run_async_wasm_package(&dir, "moon/async_timer_workspace/main"),
-        expect![[r#"
-            timer resumed
-            Total tests: 1, passed: 1, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_workspace_fs_smoke() {
-    let dir = TestDir::new("moon_test/async_wasm_workspace_fs");
-    prepare_async_wasm_workspace(&dir);
-
-    check(
-        run_async_wasm_package(&dir, "moon/async_fs_workspace/main"),
-        expect![[r#"
-            Total tests: 2, passed: 2, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_src_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async"),
-        expect![[r#"
-            Total tests: 91, passed: 91, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_aqueue_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/aqueue"),
-        expect![[r#"
-            Total tests: 52, passed: 52, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_cond_var_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/cond_var"),
-        expect![[r#"
-            Total tests: 8, passed: 8, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_semaphore_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/semaphore"),
-        expect![[r#"
-            Total tests: 12, passed: 12, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_fs_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/fs"),
-        expect![[r#"
-        Total tests: 39, passed: 39, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_pipe_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/pipe"),
-        expect![[r#"
-            Total tests: 3, passed: 3, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_stdio_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/stdio"),
-        expect![[r#"
-            Total tests: 1, passed: 1, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_socket_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/socket"),
-        expect![[r#"
-            Total tests: 24, passed: 24, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_tls_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/tls"),
-        expect![[r#"
-            Total tests: 7, passed: 7, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_http_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/http"),
-        expect![[r#"
-            Total tests: 49, passed: 49, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_process_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/process"),
-        expect![[r#"
-            Total tests: 32, passed: 32, failed: 0.
-        "#]],
-    );
-}
-
-#[test]
-fn test_async_wasm_upstream_signal_package() {
-    check(
-        run_upstream_async_wasm_package("moonbitlang/async/signal"),
-        expect![[r#"
-            Total tests: 3, passed: 3, failed: 0.
-        "#]],
-    );
 }
 
 #[test]
