+++ docs/dev/reference/testing-strategy.md
@@ -461,6 +461,10 @@ These should not be mixed into the default fast path if they can be isolated:
 
 - `third_party`
 - `test_moonbitlang_x`
+- the full upstream `moonbitlang/async` wasm conformance test in
+  `moon_test::async_wasm`;
+  its dedicated CI workflow runs for Moonrun or upstream async changes and on
+  scheduled builds
 - network-dependent parts of `moon_coverage`
 - git-install tests in `crates/moon/tests/test_cases/mod.rs`
 - environment-dependent parts of `moon_version`
